added new "revert" function
basicaly undoing deleting of original gif file

+++ bot.py
@@ -18,8 +18,12 @@ async def on_ready():
 @bot.event
 async def on_message(message):
 	for attachment in message.attachments:
-		if attachment['filename'].split('.')[-1] == 'gif':
+		if attachment['filename'].split('.')[-1] == 'gif' and message.author != bot.user:
 			bot.delete_message(message)
+			try:
+				os.remove (message.channel.id + '.gif')
+			except FileNotFoundError as e:
+				print(str(e) + " on channel " + message.channel.id +". first gif on this channel?")
 			await bot.send_typing(message.channel)
 			await bot.change_presence(game=discord.Game(name='downloading gif'))
 			gifsize = attachment['size']
@@ -31,13 +35,12 @@ async def on_message(message):
         			'User-Agent': 'Mozilla/5.0 (Macintosh; Intel Mac OS X 10_9_3) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/35.0.1916.47 Safari/537.36'
     			}
 			)
-			with urllib.request.urlopen(req) as response, open('temp.gif', 'wb') as out_file:
+			with urllib.request.urlopen(req) as response, open(message.channel.id + '.gif', 'wb') as out_file:
 				shutil.copyfileobj(response, out_file)
 			await bot.change_presence(game=discord.Game(name='compressing gif'))
-			clip = moviepy.editor.VideoFileClip ('temp.gif')
+			clip = moviepy.editor.VideoFileClip (message.channel.id + '.gif')
 			clip.write_videofile('temp.webm')
 			webmsize = os.path.getsize('temp.webm')
-			os.remove ('temp.gif')
 			await bot.change_presence(game=discord.Game(name='done!'))
 			await bot.send_file(message.channel, 'temp.webm')
 			await bot.send_message(message.channel, 'original sent by ' + message.author.name)
@@ -46,5 +49,10 @@ async def on_message(message):
 			await asyncio.sleep(5)
 			await bot.change_presence(game=discord.Game(name='waiting for gif'))
 			os.remove ('temp.webm')
+	if message.content == "!revert":
+		await bot.send_message(message.channel, "reverting last gif compressed in this channel...")
+		await bot.send_file(message.channel, message.channel.id + ".gif")
+	if message.content == "!gifhelp":
+		await bot.send_message(message.channel, "```avable commends\n!revert```")
 
 bot.run(token)
